refactor(analytics): extract parseDate and parseLimit to analytics-utils
Move the duplicated parseDate() and parseLimit() helpers out of both
analytics controllers into src/analytics/common/analytics-utils.ts,
alongside the existing safeAvg and k-anonymity constants.

- parseDate: validates ISO 8601 strings, throws BadRequestException on
  invalid input; was identical in both controllers
- parseLimit: strict digit-only regex guard + 1–100 range check; was
  identical in both controllers after the parseInt("5xyz") fix
